refactored file strucrure at project
