--- src/index.css
@@ -1,5 +1,0 @@
-*{
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
