+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import {BrowserRouter} from "react-router-dom";
-import './index.css'
+import './assets/stylesheets/index.css'
 import Routes from './routes'
 import Header from "./components/Header";
 import Footer from "./components/Footer";
